test: move standalone import.meta tests to configCases
Move comprehensive standalone import.meta tests (variable assignment,
return value, argument passing, object identity, runtime properties)
from test/cases/ to test/configCases/parsing/import-meta-standalone/
with ESM output config to avoid SyntaxError in non-module test suites.
